Allow user to restrict KDC to specific addresses
krb5kdc has always only supported binding to the wildcard addresses.
Two configuration options called kdc_listen and kdc_tcp_listen,
which replace the options kdc_ports and kdc_tcp_ports, have been
added that allow specifying a list of ports or addresses with
optional ports.  The configuration option expects all specified
addresses to be local and be able to be bound to.  If any of them
fail then krb5kdc will fail to start.  This means that if kdc.conf
is copied from the master to the slaves then it needs to be updated
for each individual slave's interfaces.

Technically hostnames can be specified, though this will complicate
things if a DNS query has to be done to find the hostname, but any
issues that would arise from that would be a configuration problem.

ticket: 1093
ticket: 6742