Remove port 750 from the KDC default ports
The KDC was still listening on port 750 despite the fact that
this functionality was supposed to have been removed in the
past.  Remove port 750 from the lists of UDP ports that the KDC
listens on.  Also, remove the port 750 from the default ports
that the client connect to, and from example config fragments.