Remove workaround when binding to wildcard udp addresses without pktinfo.
Currently if the system doesn't support pktinfo and kadmind or krb5kdc
are binding to a upd address the system krb5 binds to all the local
addresses. Currently most systems support pktinfo, so the workaround
isn't really required anymore. Removing the workaround will only affect
systems that don't have pktinfo support and have multiple NICs and are
listening on a wildcard address. Users with this setup may specify each
of the local addresses directly that kadmind or krb5kdc should listen on
in kdc.conf.

ticket: 8348