Use LSA cache as default when a TGT cannot be retrieved from LSA cache
When UAC is enabled and a domain user with Administrator privileges logs
in the TGT is inaccessible.  This is because if the TGT is accessible it
may be possible for a non-elevated user session to get a ticket with an
unfiltered PAC and bypass the UAC.  This causes a bug where the current
tickets in the LSA cache are all copied to the ccache except the TGT,
which prevents a user session from getting new tickets.  In this case,
ms2mit will now set the LSA cache to the default cache to workaround
this issue.