Add new DAL function for renaming principals
Previously libkadm5srv renamed principals by getting the principal
entry, renaming the entry, putting it in the DB, then deleting the
old one.  This does not work in certain plugins such as LDAP.  A
new DAL function is necessary to support all plugins.  A default
rename function was added to libkdb that performs the previous
functionality of adding and deleting the principal entry.