Implement principal renaming in LDAP
The previous way of renaming principals by adding a new entry and
deleting the old one did not work in LDAP due to issues such as
requiring renaming the DN in standalone principals, and modifying
the DN in place for non-standalone principals.  Add a function to
libkdb_ldap that implements the DAL rename_principal function that
properly renames the DN and attributes when necessary.

ticket: 8065